split clit types to own file
