+++ harpy/align.py
@@ -9,7 +9,8 @@ from rich import box
 from rich.table import Table
 import rich_click as click
 from ._conda import create_conda_recipes
-from ._misc import fetch_report, fetch_rule, snakemake_log, ContigList
+from ._misc import fetch_report, fetch_rule, snakemake_log
+from ._cli_types import  ContigList
 from ._launch import launch_snakemake
 from ._parsers import parse_fastq_inputs
 from ._printing import print_error, print_solution, print_notice
