+++ harpy/snakefiles/metassembly.smk
@@ -22,12 +22,12 @@ extra = config["spades"].get("extra", "")
 spadesdir = f"{outdir}/{'cloudspades' if not ignore_bx else 'spades'}_assembly"
 skip_reports  = config["reports"]["skip"]
 organism = config["reports"]["organism_type"]
-if organism == "eukaryote":
-    lineagedb = "eukaryota"
-elif organism == "fungus":
-    lineagedb = "fungi"
-else:
-    lineagedb = "bacteria"
+lineage_map = {
+    "eukaryote": "eukaryota",
+    "fungus": "fungi",
+    "bacteria": "bacteria"
+}
+lineagedb = lineage_map.get(organism, "bacteria")
 
 rule sort_by_barcode:
     input:
